Refactor Task to have State
